# [Attention] Optimize make_local_attention_virtual_batches for Flash Attention