## Purpose
This PR improves efficiency of `cu_seqlens_q_local` and `batch_indices` in `make_local_attention_virtual_batches`.

For `cu_seqlens_q_local`, original code creates multiple intermediate arrays (`np.cumsum`, `np.pad`, `.astype(np.int32)`); also `np.pad` is not necessary. In the new code, we pre-allocate output array and let `np.cumsum` directly write to the pre-allocated array.

For `batch_indices`, used the following operations to avoid potential copies and intermediate arrays:
- Uses broadcasting with `block_starts[:, None]` instead of `np.broadcast_to()` and `np.expand_dims()`
- Uses `reshape(-1)` instead of `flatten()`

## Test Plan
Injected timing code to measure latency improvement, and used loadgen with input_length=2048 and output_length=256

Verified correctness using `gsm8k` eval.

Llama 4:
`cu_seqlens_q_local` consistently improved from 35μs to 6μs
`batch_indices` consistently improved from 53μs to 34μs



## Test Result

## (Optional) Documentation Update

---
<details>
<summary> Essential Elements of an Effective PR Description Checklist </summary>

- [x] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [x] The test plan, such as providing test command.
- [x] The test results, such as pasting the results comparison before and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.
</details>